--- app/src/main/java/com/example/swapit/ui/shopping/ShoppingViewModelFactory.kt
@@ -1,17 +1,0 @@
-package com.example.swapit.ui.shopping
-
-import ShoppingViewModel
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.example.swapit.domain.repository.ShoppingRepository
-
-class ShoppingViewModelFactory(private val repository: ShoppingRepository) :
-    ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(ShoppingViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return ShoppingViewModel(repository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
