+++ app/src/main/java/com/example/swapit/ui/shopping/model/CalculateTime.kt
@@ -1,4 +1,4 @@
-package com.example.swapit.ui.shopping
+package com.example.swapit.ui.shopping.model
 
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
